gcp-chroma-cloudrun/cloudrun/requirements.txt
streamlit
google-cloud-storage
langchain==0.1.14
chromadb==0.4.24
langchain_google_vertexai==0.1.2
langchain-google-community
openai